requirements.txt
altgraph>=0.17.4
astroid>=3.3.9
beautifulsoup4>=4.13.3
black>=25.1.0
bs4>=0.0.2
certifi>=2025.1.31
cffi>=1.17.1
charset-normalizer>=3.4.1
click>=8.1.8
colorama>=0.4.6
cryptography>=44.0.2
dill>=0.4.0
et_xmlfile>=2.0.0
flake8>=7.2.0
idna>=3.10
imageio>=2.37.0
iniconfig>=2.1.0
isort>=6.0.1
mccabe>=0.7.0
mypy-extensions>=1.0.0
Nuitka>=2.6.8
numpy>=2.2.3
openpyxl>=3.1.5
ordered-set>=4.1.0
packaging>=24.2
pandas>=2.2.3
pathspec>=0.12.1
pdfminer>=20191125
pdfminer.six>=20240706
pefile>=2023.2.7
pillow>=11.1.0
platformdirs>=4.3.7
pluggy>=1.5.0
psutil>=7.0.0
pycodestyle>=2.13.0
pycparser>=2.22
pycryptodome>=3.21.0
pyflakes>=3.3.2
pyinstaller>=6.12.0
pyinstaller-hooks-contrib>=2025.1
pylint>=3.3.6
PyPDF2>=3.0.1
pytest>=8.3.5
python-dateutil>=2.9.0.post0
pytz>=2025.2
pywin32-ctypes>=0.2.3
requests>=2.32.3
setuptools>=76.0.0
six>=1.17.0
soupsieve>=2.6
tomlkit>=0.13.2
typing_extensions>=4.12.2
tzdata>=2025.2
urllib3>=2.3.0
zstandard>=0.23.0